+++ LeProjet/LeProjet/files/src/FlyingBomb.h
@@ -4,7 +4,7 @@
 class FlyingBomb : public Enemy
 {
 public:
-	FlyingBomb(int x, int y, int dep, int arr);
+	FlyingBomb(float x, float y, float dep, float arr);
 
 	void DrawUnit() override;
 	int UpdateUnit(EnvItem* envItems, size_t envItemsLength, float delta) override;
